+++ packages/core/src/types.ts
@@ -7,7 +7,7 @@ import type { Context as ElysiaContext } from "elysia";
 import type { FastifyReply, FastifyRequest } from "fastify";
 import type { H3Event } from "h3";
 import type { Context as HonoContext } from "hono";
-import type { universalSymbol } from "./const";
+import type { methodSymbol, nameSymbol, optionsToSymbols, orderSymbol, pathSymbol, universalSymbol } from "./const";
 
 // Helpers
 
@@ -235,3 +235,69 @@ export type UniversalHandler<InContext extends Universal.Context = Universal.Con
 ) => Awaitable<Response>;
 
 export type Get<T extends unknown[], U> = (...args: T) => U;
+
+// Router
+
+export enum MiddlewareOrder {
+  // Pre-handler Middlewares
+  GUARD = -1000, // Guard middleware: Ensures specific conditions or headers are met before proceeding.
+  AUTHENTICATION = -900, // Authentication middleware: Verifies user credentials or tokens.
+  AUTHORIZATION = -800, // Authorization middleware: Ensures the user has permissions for the route.
+  RATE_LIMITING = -700, // Rate limiting middleware: Prevents excessive requests from a client.
+  INPUT_VALIDATION = -600, // Input validation middleware: Validates the request payload or query parameters.
+  CORS = -500, // CORS middleware: Handles Cross-Origin Resource Sharing settings.
+  PARSING = -400, // Parsing middleware: Parses body payloads (e.g., JSON, URL-encoded, multipart).
+  CUSTOM_PRE_PROCESSING = -300, // Custom pre-processing middleware: Any custom logic before the main handler.
+
+  // Main Handler
+  HANDLER = 0, // Main handler that generates the response.
+
+  // Post-handler Middlewares
+  RESPONSE_TRANSFORM = 100, // Response transformation middleware: Modifies the response payload.
+  HEADER_MANAGEMENT = 200, // Header management middleware: Adds or modifies HTTP headers (e.g., caching, content type).
+  RESPONSE_COMPRESSION = 300, // Response compression middleware: Compresses the response payload (e.g., gzip, brotli).
+  RESPONSE_CACHING = 400, // Response caching middleware: Implements caching strategies (e.g., ETag, cache-control).
+  LOGGING = 500, // Logging middleware: Logs request and response information.
+  ERROR_HANDLING = 600, // Error handling middleware: Processes errors and returns user-friendly responses.
+  CUSTOM_POST_PROCESSING = 700, // Custom post-processing middleware: Any custom logic after the response is generated.
+}
+
+export interface UniversalSymbols {
+  [nameSymbol]: string;
+  [methodSymbol]: HttpMethod | HttpMethod[];
+  [pathSymbol]: string;
+  [orderSymbol]: MiddlewareOrder | number;
+}
+
+type OptionsToSymbols = typeof optionsToSymbols;
+
+export type UniversalOptions = {
+  [K in keyof OptionsToSymbols]: UniversalSymbols[OptionsToSymbols[K]];
+};
+
+export interface UniversalOptionsArg extends Partial<UniversalOptions> {
+  /**
+   * @default true
+   */
+  immutable?: boolean;
+}
+
+export type WithUniversalSymbols<T extends UniversalOptionsArg> = Pick<
+  UniversalSymbols,
+  OptionsToSymbols[keyof T & keyof OptionsToSymbols]
+>;
+
+// https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods
+export type HttpMethod = "GET" | "HEAD" | "POST" | "PUT" | "DELETE" | "CONNECT" | "OPTIONS" | "TRACE" | "PATCH";
+export type Decorate<T> = T & Partial<UniversalSymbols>;
+
+export type DecoratedMiddleware =
+  | Decorate<UniversalMiddleware>
+  | { [universalSymbol]: Decorate<UniversalMiddleware> }
+  | (Decorate<AnyFn> & { [universalSymbol]: UniversalMiddleware });
+
+export interface UniversalRouterInterface {
+  use(middleware: DecoratedMiddleware): this;
+  route(handler: DecoratedMiddleware): this;
+  applyCatchAll(): void;
+}
