+++ packages/core/src/utils.ts
@@ -1,6 +1,6 @@
 import type { OutgoingHttpHeaders } from "node:http";
 import { unboundSymbol, universalSymbol } from "./const";
-import type { AnyFn, SetThis, UniversalFn, UniversalHandler, UniversalMiddleware } from "./types";
+import type { AnyFn, SetThis, UniversalFn, UniversalHandler, UniversalMiddleware, UniversalSymbols } from "./types";
 
 export function isBodyInit(value: unknown): value is BodyInit {
   return (
@@ -47,6 +47,33 @@ function normalizeHttpHeader(value: string | string[] | number | undefined): str
     return value.join(", ");
   }
   return (value as string) || "";
+}
+
+export function getUniversal<T extends object>(subject: T | { [universalSymbol]: T }): T {
+  return universalSymbol in subject ? subject[universalSymbol] : subject;
+}
+
+export function getUniversalProp<T extends object, K extends keyof UniversalSymbols>(
+  subject: T | { [universalSymbol]: T },
+  prop: K,
+): UniversalSymbols[K] | undefined {
+  if (prop in subject) return (subject as UniversalSymbols)[prop];
+  if (universalSymbol in subject) return (subject as Record<symbol, UniversalSymbols>)[universalSymbol][prop];
+  return undefined;
+}
+
+/**
+ * @internal
+ */
+export function cloneFunction<F extends AnyFn>(originalFn: F): F {
+  const extendedFunction = function (this: unknown, ...args: unknown[]) {
+    // Use Reflect.apply to invoke the original function
+    return Reflect.apply(Object.getPrototypeOf(extendedFunction), this, args);
+  };
+
+  Object.setPrototypeOf(extendedFunction, originalFn);
+
+  return extendedFunction as F;
 }
 
 /**
